refactor(capabilities): migrate the fs cap to the struct-hosted actor macro
Apply the ADR-0123 struct-hosted transform to the fs cap (#2379).
